pyproject.toml
[project]
name = "lmbatbot"
version = "0.1.2"
description = "A simple utility Bot with some fun commands"
readme = "README.md"
authors = [
    { name = "Lorenzo Bevilacqua", email = "43483037+ardubev16@users.noreply.github.com" },
]
requires-python = ">=3.12"
dependencies = [
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "python-telegram-bot==21.9",
    "requests==2.32.3",
    "sqlalchemy==2.0.36",
]

[dependency-groups]
dev = ["ruff==0.8.4", "vulture==2.14"]

[project.scripts]
lmbatbot = "lmbatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "strict"
reportUnknownMemberType = false
reportUnknownVariableType = false

[tool.ruff]
extend = "./sync-config/pyproject.toml"

[tool.ruff.lint]
ignore = [
    "S101", # Checks for uses of the assert keyword.
]
